+++ plugins/css.js
@@ -1,5 +1,4 @@
 import textLoader from "../loaders/text.js";
-import { join } from "../deps/path.js";
 import { Coder, ImportPlugin, parse, Tasks } from "../deps/stylecow.js";
 import { postcss, postcssPresetEnv } from "../deps/postcss.js";
 
@@ -20,8 +19,8 @@ export default function () {
     site.process([".css"], processor);
 
     async function processor(page) {
-      const from = join(site.options.src, page.src.path + page.src.ext);
-      const to = join(site.options.dest, page.dest.path + page.dest.ext);
+      const from = site.src(page.src.path + page.src.ext);
+      const to = site.dest(page.dest.path + page.dest.ext);
 
       //Resolve @import with stylecow
       const css = parse(page.content, "Root", null, from);