+++ CHANGELOG.md
@@ -5,14 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.8.0] - Unreleased
+## [0.8.0] - 2020-10-27
 ### Added
 - New `loadAssets()` to register assets loaders
+- New argument in CLI to build the site in a different directory and even choose a different _config.js file.
 
 ### Changed
 - Renamed `load()` to `loadPages()` and removed the `asset` argument.
 - Renamed `data()` to `loadData()`
 - Renamed `helper()` to `data()`
+- Updating files (in a watch process) dispatches the events `beforeUpdate` and `afterUpdate` (instead `beforeBuild` and `afterBuild`)
 
 ## [0.7.3] - 2020-10-17
 ### Changed
@@ -139,7 +141,7 @@ First version
 
 [#1]: https://github.com/oscarotero/lume/issues/1
 
-[0.8.0]: https://github.com/oscarotero/lume/compare/v0.7.3...HEAD
+[0.8.0]: https://github.com/oscarotero/lume/compare/v0.7.3...v0.8.0
 [0.7.3]: https://github.com/oscarotero/lume/compare/v0.7.2...v0.7.3
 [0.7.2]: https://github.com/oscarotero/lume/compare/v0.7.1...v0.7.2
 [0.7.1]: https://github.com/oscarotero/lume/compare/v0.7.0...v0.7.1